+++ src/projects/universityAssistant/Services/UniversityService/Core/UniversityService.Domain/Entities/Provience.cs
@@ -4,8 +4,8 @@ namespace UniversityService.Domain.Entities;
 
 public class Provience : Entity
 {
-    public long Id { get; set; }
-    public long CountryId { get; set; }
+    public int Id { get; set; }
+    public int CountryId { get; set; }
     public string Name { get; set; }
     public virtual Country Country { get; set; }
 }
